Extract body related actions to a trait.
